S06P31E106-97 Fix: 회원정보 예외처리 수정

+++ backend/src/main/java/com/ssafy/alpaca/api/service/UserService.java
@@ -14,6 +14,9 @@ import com.ssafy.alpaca.db.repository.LogoutAccessTokenRedisRepository;
 import com.ssafy.alpaca.db.repository.RefreshTokenRedisRepository;
 import com.ssafy.alpaca.db.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -39,6 +42,12 @@ public class UserService {
         Map<String, String> map = new HashMap<>();
         map.put("message", returnMessage);
         return map;
+    }
+
+    private String getCurrentUsername() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        UserDetails principal = (UserDetails) authentication.getPrincipal();
+        return principal.getUsername();
     }
 
     public Map<String, String> checkUsername(String username) {
@@ -168,12 +177,16 @@ public class UserService {
     }
 
     // 아래부터 UserController
-    public Map<String, String> updateUser(String id, UserUpdateReq userUpdateReq) {
+    public Map<String, String> updateUser(String id, UserUpdateReq userUpdateReq) throws IllegalAccessException {
         Optional<User> user = userRepository.findById(id);
         if (user.isEmpty()) {
             throw new NoSuchElementException(ExceptionUtil.USER_NOT_FOUND);
         }
+        String username = getCurrentUsername();
         User changedUser = user.get();
+        if (!changedUser.getUsername().equals(username)) {
+            throw new IllegalAccessException(ExceptionUtil.NOT_MYSELF);
+        }
         changedUser.setNickname(userUpdateReq.getNickname());
         changedUser.setInfo(userUpdateReq.getInfo());
         changedUser.setTheme(userUpdateReq.getTheme());
@@ -182,12 +195,16 @@ public class UserService {
         return getMessage("성공적으로 수정되었습니다.");
     }
 
-    public void deleteUser(String id) {
+    public void deleteUser(String id) throws IllegalAccessException {
         Optional<User> user = userRepository.findById(id);
         if (user.isEmpty()) {
             throw new NoSuchElementException(ExceptionUtil.USER_NOT_FOUND);
         }
+        String username = getCurrentUsername();
         User deleteUser = user.get();
+        if (!deleteUser.getUsername().equals(username)) {
+            throw new IllegalAccessException(ExceptionUtil.NOT_MYSELF);
+        }
         userRepository.delete(deleteUser);
     }
